requirements-ai-control.txt
-r ai_agents/requirements.txt
fastapi==0.135.1
prometheus-client==0.24.1
docker==7.1.0
uvicorn==0.37.0
PyYAML==6.0.3
Jinja2==3.1.6
pydantic-settings==2.10.1
